chore: update ignore rules to include certain directories and exclude their builds
The changes add rules to the `.gitignore` file, making build directories in certain locations not be ignored. The directories are `gradle/wrapper`, `src/main` and `src/test` in all folders, and `bin` in `src/main` and `src/test`. The purpose of this seems to be to include the locally-built artifacts in the source directories in the code repository. Some new ignore rules are added for popular code editors like IntelliJ IDEA and Eclipse. Certain files in the project structure related to testing and build processes are still ignored.