chore: update Gradle task configuration for running tests and inspections
The Gradle task configuration for running tests and inspections was updated in `.run/Run IDE for UI Tests.run.xml` and `.run/Run Qodana.run.xml`. Specifically, the `scriptParameters` option was empty during tests and inspections runs and was updated with specific values for `runIdeForUiTests`, `cleanInspections`, and `runInspections` tasks. The `RunAsTest` flag was also set to `false`.